move cloud init datastructures over to their own file, the naming in here is terrible
